requirements.txt
pyautogui
keyboard
pillow
mss
colorama